Fix SCMStepTest#scmVars for PCT
Lots of our dependencies depended on old, stale versions of
workflow-scm-step, so we had to exclude them all over the place. Then
we had to bump svnkit-cli to match the subversion plugin version, but
at least on my Mac, going to 1.9.1 hung due to "svn: E204899: Can't
get exclusive lock on file ...". Joy! Bumping again to 1.9.2 and
forcing svnkit proper to the same version fixed things.